Moved all interfaces to UrlShortenerInterfaces.ts
